dtbs: don't fail when no dtbs are built
In an arm64 tinyconfig build, the dtbs and dtbs_install targets run but
don't do anything, causing the tar call at the end to fail.
Unfortunately I could not discover yet a way of checking if the dtbs
would have been built, in which case we could skip the whole thing.
CONFIG_OF_EARLY_FLATTREE, which is the main config that makes the `all`
target depend on `dtbs`, is also set to `y` in a tinyconfig build.

Creating the directory beforehand causes the tarball to be sucessfully
created, even though it's empty (which is expected since the user
requested a tinyconfig build, which does not build dtbs).

Closes: #47