Rename to --docker-image/$TUXMAKE_DOCKER_IMAGE to --image/$TUXMAKE_IMAGE
Given that we have now more runtimes than docker that implement the
concept of images, it no longer makes sense to have a docker-specific
command line argument. Thus, we rename --docker-image to --image, and
the $TUXMAKE_DOCKER_IMAGE environment variable is now $TUXMAKE_IMAGE.

The old names are still accepted for backwards compatibility, but will
be removed in the future.

Additionally, --image will now only imply --runtime=docker if no runtime
is explicitly specified in the command line.