Parallelize docker image builds
Each buildjob only handles the images for the same toolchain, e.g.
gcc-10, or clang-nightly.

The gitlab ci configuration is autogenerated based on the metadata in
docker.ini. It needs to receive as input a TUXMAKE_DOCKER_IMAGES
variable, which can be either "monthly" or "daily", and will
enable/disable jobs based on the rebuild periodicity of each toolchain.