Build: handle missing artifacts
If an expected artifact is not in fact in place after the build, tuxmake
would crash trying to copy it to the output directory.

Instead, if the expected artifact is not found, mark the build of the
corresponding target as failed. This way copying the artifact will be
skipped.

Fixes #96