metadata: kernelversion: avoid all output from make
When building a Debian package out of tuxmake, the kernelversion test
fails because the output has the "Entering directory foo" and "Leaving
directory foo" messages.

Note that --no-print-directory is a GNU make extension, so this might
not work on othe make implementations. OTOH we don't expect tuxmake to
build on runtimes without GNU make being the default make.