cmdline: bash_completion: fix completion of directory names
For some reason the previous version would stop completion at the first
directory, i.e. `tuxmake -o /tm<TAB>` would complete to /tmp and not
allow me to pick a directory inside /tmp (by completing to /tmp/ and not
adding a space). I tested several variations by trial and error, and
these changes are the ones that make it more or less work.

There is still a problem in which it will still complete file (i.e. no
directory) names, but the current state is better than the previous
where you could not correctly pick a subdirectory.