target: add kconfig_add target option
Allow a target config to specify kconfig fragments with the kconfig_add
option. The syntax is the same as the --kconfig-add CLI option. A single
option per target can specify multiple fragments, separated by space.
Quotes may be used if necessary:

        kconfig_add = CONFIG_KPROBES=y
                      "# CONFIG_DEBUG_INFO_BTF is not set"
                      CONFIG_TRACE=y

After cleaning up the list of targets, the Build class extracts the
kconfig_add options of each target and adds it to the kconfig_add list
specified on the command-line. Then the config target processes those
options as usual.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>