wrapper: split preparation in pre- and post- runtime preparation
Now that a single container is started at the beginning of the build and
not with each command, a part of the wrapper preparation must be done
before the runtime preparation, e.g. creating the necessary directories
on the host system so that they can be bind mounted in the containers.