docker: install most development libraries from the default release
This avoids, for example, pulling libssl-dev from experimental when also
pulling gcc-11 from experimental. Only the packages closely related to
to the toolchain such as libc6 and binutils need to be installed using
any provided install options to make sure they are coinstallable with
the newer toolchain.

This unbreaks building gcc-11 images.

Closes #126