fakelinux: implement LLVM=1 more faithfully
Also set the linker to ld.lld, so that we can actually simulate LLVM=1
builds. In test/conftest.py, fake_compilers needs to be adjusted to also
create a fake linker. A couple of integration tests need to be skipped
if ld.lld is not available (e.g. Ubuntu 18.08 on arm64).