tuxmake-check-environment: make full cross compilation optional
While cross compilation of user space code is supported in Debian, it's
not so readily available in other distributions. So make failures to
cross compile binaries against a libc non-fatal by default, and allow
them to be made fatal with a command line argument.

Then, change the tuxmake image build process to make those failures
fatal on the fully supported architectures, instead of making it
non-fatal on the other ones.