metadata: extract build hardware information
For extracting the free disk space before the build, we need to change
the life cycle of the metadata extractor object inside the build object.
It's now instantiated together with the build, and gained a
before_build() method that gets called, well, before the build.

Also add the ability to implement more complex metadata extrators as
Python classes, the first case being FreeDiskSpace.

Fixes: https://gitlab.com/Linaro/tuxmake/-/issues/142