metadata: fix KeyError when running Python extractor
This is not directly possible with the current code, but will become an
issue if/when we have a metadata group that contains only Python
extractors (i.e. no [commands] section). It was already seen happening
due to some corner case, but so far we could not reproduce it.

This also adds a dependency injection point into MetadataCollector for
setting a different set of metadata handlers than the default (i.e. all
of the defined ones).

Reported-By: Anders Roxell <anders.roxell@linaro.org>
Tested-By: Anders Roxell <anders.roxell@linaro.org>