build, cmdline: add "fail fast" feature
This will cause tuxmake to abort the build as soon as there is any
errors during the build. This means not only not telling make to "keep
going", but also stopping the build after any target fails, even if
later targets don't depend on it and could normally be built.