target: default: run after kselftest-merge
Ideally, we would want `kselftest-merge` to declare that it needs to run
before `default`, but that would involve adding a "runs_before"
mechanism to the target dependency handling, and that would require more
thought to be put into the issue. For now, let's just to the simplest
thing that works, and maybe revisit that in the future if other
instances of this show up.

Fixes: https://gitlab.com/Linaro/tuxmake/-/issues/156