images: fix installation of packages on native images
When building the arm64 build images on arm64 hosts, the extra packages
for arm-linux-gnueabihf where not being installed because they were
being handled together with the "normal" cross compilers for arm64,
which don't need (and can't) be installed because gcc-aarch64-linux-gnu
is not available on arm64 since that is provided by the regular gcc
package.

Stop skipping the installation in the Dockerfile when the host
architecture is different from the native one, and instead compose the
list of packages to install differently based on that, and install
whatever list is obtained unconditionally.

Fixes: 2d714fb37e03 ("images: ensure 32-bit ARM toolchain is included in native arm64")