test/integration: Look for Arch Linux's shunit2
shunit2 is not packaged in the official Arch Linux repo but it is
available in the Arch Linux User Repository. Unfortunately, the path
that it is installed to is different than Debian or Fedora. To avoid
forking the shunit2 PKGBUILD to change the path, just check for the
existing path first and use it if it exists.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>