ci: Wire up Arch Linux package building
Add a PKGBUILD for tuxmake and build it as a part of CI. The unit tests
are automatically run via check().

The Arch Linux container requires a little bit more setup than the
Debian or Fedora containers:

1. By default, the Arch Linux container does not contain all of the
   locales available through glibc. Remove the "NoExtract" lines from
   /etc/pacman.conf and reinstall glibc so that C.UTF-8 is available, as
   that is what the CI environment expects.

2. The root user is not allowed to build packages with makepkg for
   security and safety reasons; add a "build" user to handle this.

3. Because the repository is technically owned by root in the container,
   we need to add the repository folder to the "safe" paths for git,
   otherwise the build user cannot access it, due to the hardening added
   by the solution for CVE-2022-24765.

The integration tests are not run as part of this change, as there are a
few quirks about Arch Linux that need to be handled first.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>