docker: include native versions of libraries
The kselftest build system currently calls `pkg-config` from PATH even
when cross compiling. On Debian, there are ${arch}-pkg-config wrapper
scripts that will call pkg-config pointing to the correct directories
when cross-compiling, so architecture-specific flags are set correctly.

This is a workaround for building kselftest before the build system is
improved. Having the native version of the libraries installed will
cause the pkg-config calls to not fail, while having the foreign
libraries will cause the compilation to actually work.

Note that this could still go wrong if pkg-config is used to get
compilatino flags for a library that has architecture-specific flags,
for example, glib and libpci:

    $ pkg-config --cflags glib-2.0
    -I/usr/include/glib-2.0 -I/usr/lib/x86_64-linux-gnu/glib-2.0/include
    $ pkg-config --cflags libpci
    -I/usr/include/x86_64-linux-gnu

Fixes: https://gitlab.com/Linaro/tuxmake/-/issues/179