docker: clang-nightly: build upon Debian testing
The last upload of linux to bullseye backports at the time of writing
this failed to build on s390x and a few other architectures, causing
linux-libc-dev to not be cross-installable and thus failing our builds.
Given the most important bits are the LLVM toolchain anyway, I think
it's OK to upgrade the OS like this.