test: test_utils: Modify environment rather than replace it
TestGetDirectoryTimestamp.test_git() replaces the subprocess environment
when calling "git commit", in order to set the GIT_COMMITTER_DATE variable.

This causes a problem when building in GNU Guix, because the distribution
installs binaries in non-standard locations (there's no /bin or /usr/bin
there) and relies on PATH pointing to where the available programs are
installed.

Without PATH in the environment, the test can't find the git binary and
thus the test fails. I didn't check, but I believe Nix would have the same
issue.

To fix the problem, change the test to add the desired variable to the
environment instead of replacing it.

Signed-off-by: Thiago Jung Bauermann <thiago.bauermann@linaro.org>