runtime: containers: protect source directory from writing
To ensures that no writes will be made to the underlying source
directory, we mount the source directory inside the container as
a copy-on-write overlay.

The linux build will usually not write to the source directory when
doing as out-of-tree build (O=), which is the case for all tuxmake
builds. Other parts of tree will, however, due to currently existing
bugs in their build system (e.g. kselftest as of writing this).

Fixes: https://gitlab.com/Linaro/tuxmake/-/issues/181