cmdline: Fix spurious --download-all-korg-gcc-toolchains in reproduced commands
The --download-all-korg-gcc-toolchains flag was incorrectly appearing in
reproduced command lines generated by the CommandLine.reproduce()
method.  This flag is an action option for downloading toolchains, not a
build parameter that should be included when reproducing builds.

Fixed by adding "download_all_korg_gcc_toolchains" to the ignore list in
the CommandLine class, preventing it from being included in reproduced
command lines. Updated the existing parametrized test to verify this
flag is properly ignored.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>